Closes #1; Fixed collision animation frames.
The issue was caused during the comparison of the bounded velocity of "0" with a negative target velocity. Basically I just check to see if the signs are different to determine whether or not to put the character in the "pivoting" state. Added a quick fix up top to always use the "STILL" state when the player is at either extremity of the screen (left or right).